fix: consolidate spark catalog operations to connection level
